Refactor new search interface into separate modules
- there was no reason to have it in one giant module
- we already had it split into implied sections in comments, so split into modules matching those
