+++ src/search/index.js
@@ -79,18 +79,13 @@ export async function handleBookmarkCreation(...args) {
 //
 // Utilities
 //
-export function initSingleLookup(...outterArgs) {
-    let singleLookup
-    return async function(...args) {
-        if (!singleLookup) {
-            singleLookup = (await getBackend()).initSingleLookup(...outterArgs)
-        }
-        return await singleLookup(...args)
-    }
-}
 
 export async function grabExistingKeys(...args) {
     return await (await getBackend()).grabExistingKeys(...args)
+}
+
+export async function getPage(url) {
+    return await (await getBackend()).getPage(url)
 }
 
 //
@@ -113,8 +108,4 @@ export const indexQueue = {
     clear: async () => {
         ;(await getBackend()).indexQueue.clear()
     },
-}
-
-export async function getPage(url) {
-    return await (await getBackend()).getPage(url)
 }
