+++ src/search/search-index-new/search/index.js
@@ -1,3 +1,5 @@
+import db from '..'
+import QueryBuilder from 'src/search/query-builder'
 import { groupLatestEventsByUrl, mapUrlsToLatestEvents } from './events'
 import { mapResultsToDisplay } from './map-results-to-display'
 import { findFilteredUrls } from './filters'
@@ -21,27 +23,82 @@ import { paginate, applyScores } from './util'
  * @property {number} 1 Timestamp of latest event.
  */
 
-/**
- * Should be run from inside a Dexie transaction.
- *
- * @param {SearchParams} params
- */
-export default async function search(params) {
-    console.time('SEARCH: main search')
-    const results = await matchingPagesSearch(params)
-    console.timeEnd('SEARCH: main search')
+export async function search({ query, showOnlyBookmarks, ...restParams }) {
+    // Extract query terms via QueryBuilder (may change)
+    const qb = new QueryBuilder().searchTerm(query).get()
 
-    console.time('SEARCH: result mapping')
-    const docs = await mapResultsToDisplay(results.ids, params)
-    console.timeEnd('SEARCH: result mapping')
+    // Short-circuit search if bad term
+    if (qb.isBadTerm) {
+        return {
+            docs: [],
+            resultsExhausted: true,
+            totalCount: 0,
+            isBadTerm: true,
+        }
+    }
 
-    return { docs, totalCount: results.totalCount }
+    // Reshape needed params; prob consolidate interface later when remove old index code
+    const params = {
+        queryTerms: [...qb.query],
+        bookmarks: showOnlyBookmarks,
+        ...restParams,
+    }
+
+    const { docs, totalCount } = await db.transaction(
+        'r',
+        db.tables,
+        async () => {
+            console.time('SEARCH: main search')
+            const results = await fullSearch(params)
+            console.timeEnd('SEARCH: main search')
+
+            console.time('SEARCH: result mapping')
+            const docs = await mapResultsToDisplay(results.ids, params)
+            console.timeEnd('SEARCH: result mapping')
+
+            return { docs, totalCount: results.totalCount }
+        },
+    )
+
+    return {
+        docs,
+        resultsExhausted: docs.length < params.limit,
+        isBadTerm: qb.isBadTerm,
+        totalCount,
+    }
+}
+
+// WARNING: Inefficient; goes through entire table
+export async function getMatchingPageCount(pattern) {
+    const re = new RegExp(pattern, 'i')
+    return await db.pages.filter(page => re.test(page.url)).count()
+}
+
+export async function suggest(query = '', type, limit = 10) {
+    // Start building the WhereClause from appropriate table
+    const whereClause = (() => {
+        switch (type) {
+            case 'domain':
+                return db.pages.where('domain')
+            case 'tag':
+            default:
+                return db.tags.where('name')
+        }
+    })()
+
+    // Perform suggestion matching
+    return await whereClause
+        .startsWith(query)
+        .limit(limit)
+        .uniqueKeys()
 }
 
 /**
+ * Main search logic. Calls the rest of serach depending on input search params.
+ *
  * @param {SearchParams} params
  */
-async function matchingPagesSearch({ queryTerms = [], ...params }) {
+async function fullSearch({ queryTerms = [], ...params }) {
     const filteredUrls = await findFilteredUrls(params)
 
     let urlScoresMap
