+++ src/search/search-index-new/util.js
@@ -1,0 +1,16 @@
+import db from '.'
+import normalizeUrl from 'src/util/encode-url-for-id'
+
+export const getPage = url => db.pages.get(normalizeUrl(url))
+
+/**
+ * Hardcoded replacement for now.
+ *
+ * TODO: Maybe overhaul `import-item-creation` module to not need this (only caller)
+ */
+export async function grabExistingKeys(...args) {
+    return db.transaction('r', db.pages, db.bookmarks, async () => ({
+        histKeys: new Set(await db.pages.toCollection().primaryKeys()),
+        bmKeys: new Set(await db.bookmarks.toCollection().primaryKeys()),
+    }))
+}
